feat:풍수해 상세보기 수정 추가

+++ storage/db-core/src/main/java/com/nexsol/tpa/storage/db/core/TotalFormMemberRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
 public interface TotalFormMemberRepository
-        extends JpaRepository<TotalFormMemberEntity, Long>, JpaSpecificationExecutor<TotalFormMemberEntity> {
+        extends JpaRepository<TotalFormMemberEntity, Integer>, JpaSpecificationExecutor<TotalFormMemberEntity> {
 
     // Page<TotalFormMemberEntity> findAll(Specification<TotalFormMemberEntity> spec,
     // Pageable pageable);
